requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
black==24.2.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.3
cqlengine==0.21.0
distlib==0.3.8
Django==4.2.10
django-cors-headers==4.3.1
django-redis==5.4.0
django-safedelete==1.3.3
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==23.3.0
filelock==3.13.1
geomet==0.2.1.post1
identify==2.5.35
idna==3.6
importlib-resources==6.1.3
isort==5.13.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.12.1
pkgutil-resolve-name==1.3.10
platformdirs==4.2.0
pre-commit==3.5.0
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic-core==2.16.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
referencing==0.33.0
rpds-py==0.18.0
scylla-driver==3.26.7
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.10.0
urllib3==2.2.1
virtualenv==20.25.1
whitenoise==6.6.0
zipp==3.17.0
